Procesamiento de error en la consulta de tiempo.

+++ irosu/src/main/java/com/pixelware/model/Current.java
@@ -14,15 +14,15 @@ public class Current implements Serializable{
 
 	private static final long serialVersionUID = 1L;
 	
-	private int temp_c;
+	private double temp_c;
 	
 	public Current() {}
 
-	public int getTemp_c() {
+	public double getTemp_c() {
 		return temp_c;
 	}
 
-	public void setTemp_c(int temp_c) {
+	public void setTemp_c(double temp_c) {
 		this.temp_c = temp_c;
 	}
 }
